First release 1.0.0

+++ inc/admin/register.php
@@ -18,6 +18,7 @@ register_uninstall_hook( __FILE__, 'wp_mermaid_uninstall' );
  */
 function wp_mermaid_activation() {
 	add_option( 'wp_mermaid_js_source', 'default' );
+	add_option( 'wp_mermaid_uninstall_option', 'yes' );
 }
 
 /**
@@ -26,5 +27,10 @@ function wp_mermaid_activation() {
  * @return void
  */
 function wp_mermaid_uninstall() {
-	delete_option( 'wp_mermaid_js_source' );
+	$option_uninstall = get_option( 'wp_mermaid_uninstall_option' );
+
+	if ( 'yes' === $option_uninstall ) {
+		delete_option( 'wp_mermaid_js_source' );
+		delete_option( 'wp_mermaid_uninstall_option' );
+	}
 }
